pyproject.toml
[project]
name = "hotel-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-crum>=0.7.9",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "pytz>=2025.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.14",
]
